# apps/backend/docs/api-doc.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "مُختصِر - واجهة برمجة التطبيقات الخاصة بمُختصِر",
    "description": "هذه هي مواصفات واجهة برمجة التطبيقات لمُختصِر التي تتيح لك الوصول برمجياً إلى خدمتنا لتحويل الروابط الطويلة إلى روابط مختصرة\n\nالميزات:\n- إنشاء روابط مختصرة (مع أو بدون أسماء مستعارة مخصصة)\n- تحديث وحذف الروابط المختصرة\n- تتبع الإحصائيات الأساسية (النقرات، المرجع، إلخ)\n",
    "contact": {
      "email": "support@mukhtasar.pro"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
...
  "paths": {
...
    "/analytics/overview": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "نظرة عامة على إحصائيات الرابط",
        "description": "الحصول على نظرة عامة سريعة لإحصائيات الرابط المختصر تشمل إجمالي النقرات والزوار الفريدين",
        "operationId": "getAnalyticsOverview",
        "parameters": [
          {
            "name": "alias",
            "in": "query",
            "description": "الاسم المستعار للرابط المختصر",
            "required": true,
            "schema": {
              "type": "string",
              "example": "myalias123"
            }
          },
          {
            "name": "startDate (UTC)",
            "in": "query",
            "description": "تاريخ البداية (افتراضي: 30 يوماً مضت)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2025-08-01T00:00:00.000Z"
            }
          },
          {
            "name": "endDate (UTC)",
            "in": "query",
            "description": "تاريخ النهاية (افتراضي: الآن)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2025-08-30T23:59:59.000Z"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "تم استرداد النظرة العامة بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "total_clicks": {
                          "type": "integer",
                          "example": 1250
                        },
                        "unique_visitors": {
                          "type": "integer",
                          "example": 890
                        },
                        "active_days": {
                          "type": "integer",
                          "example": 25
                        },
                        "avg_clicks_per_day": {
                          "type": "number",
                          "example": 50.0
                        }
                      }
                    },
                    "errors": {
                      "type": "array",
                      "items": {},
                      "example": []
                    },
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "errorCode": {
                      "type": "string",
                      "example": "NO_ERROR"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitExceeded"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/analytics/browsers": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "إحصائيات المتصفحات",
        "description": "توزيع النقرات حسب نوع المتصفح",
        "operationId": "getBrowserStats",
        "parameters": [
          {
            "name": "alias",
            "in": "query",
            "description": "الاسم المستعار للرابط المختصر",
            "required": true,
            "schema": {
              "type": "string",
              "example": "myalias123"
            }
          },
          {
            "name": "startDate (UTC)",
            "in": "query",
            "description": "تاريخ البداية (افتراضي: 30 يوماً مضت)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2025-08-01T00:00:00.000Z"
            }
          },
          {
            "name": "endDate (UTC)",
            "in": "query",
            "description": "تاريخ النهاية (افتراضي: الآن)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2025-08-30T23:59:59.000Z"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "تم استرداد إحصائيات المتصفحات بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "browser": {
                            "type": "string",
                            "example": "Chrome"
                          },
                          "count": {
                            "type": "integer",
                            "example": 850
                          },
                          "percentage": {
                            "type": "number",
                            "example": 68.0
                          }
                        }
                      }
                    },
                    "errors": {
                      "type": "array",
                      "items": {},
                      "example": []
                    },
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "errorCode": {
                      "type": "string",
                      "example": "NO_ERROR"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitExceeded"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/analytics/devices": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "إحصائيات الأجهزة",
        "description": "توزيع النقرات حسب نوع الجهاز (موبايل، سطح المكتب، تابلت)",
        "operationId": "getDeviceStats",
        "parameters": [
          {
            "name": "alias",
            "in": "query",
            "description": "الاسم المستعار للرابط المختصر",
            "required": true,
            "schema": {
              "type": "string",
              "example": "myalias123"
            }
          },
          {
            "name": "startDate (UTC)",
            "in": "query",
            "description": "تاريخ البداية (افتراضي: 30 يوماً مضت)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2025-08-01T00:00:00.000Z"
            }
          },
          {
            "name": "endDate (UTC)",
            "in": "query",
            "description": "تاريخ النهاية (افتراضي: الآن)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2025-08-30T23:59:59.000Z"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "تم استرداد إحصائيات الأجهزة بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "device": {
                            "type": "string",
                            "example": "Mobile"
                          },
                          "count": {
                            "type": "integer",
                            "example": 750
                          },
                          "percentage": {
                            "type": "number",
                            "example": 60.0
                          }
                        }
                      }
                    },
                    "errors": {
                      "type": "array",
                      "items": {},
                      "example": []
                    },
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "errorCode": {
                      "type": "string",
                      "example": "NO_ERROR"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitExceeded"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/analytics/clicks-over-time": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "النقرات عبر الوقت",
        "description": "تتبع النقرات على مدار فترة زمنية معينة مع إمكانية التجميع حسب اليوم، الساعة، الأسبوع، أو الشهر",
        "operationId": "getClicksOverTime",
        "parameters": [
          {
            "name": "alias",
            "in": "query",
            "description": "الاسم المستعار للرابط المختصر",
            "required": true,
            "schema": {
              "type": "string",
              "example": "myalias123"
            }
          },
          {
            "name": "startDate (UTC)",
            "in": "query",
            "description": "تاريخ البداية (افتراضي: 90 يوماً مضت)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2025-08-01T00:00:00.000Z"
            }
          },
          {
            "name": "endDate (UTC)",
            "in": "query",
            "description": "تاريخ النهاية (افتراضي: الآن)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2025-08-30T23:59:59.000Z"
            }
          },
          {
            "name": "groupBy",
            "in": "query",
            "description": "طريقة التجميع (افتراضي: day)",
            "schema": {
              "type": "string",
              "enum": [
                "hour",
                "day",
                "week",
                "month"
              ],
              "example": "day"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "تم استرداد النقرات عبر الوقت بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "date": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-09-01"
                          },
                          "clicks": {
                            "type": "integer",
                            "example": 35
                          }
                        }
                      }
                    },
                    "errors": {
                      "type": "array",
                      "items": {},
                      "example": []
                    },
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "errorCode": {
                      "type": "string",
                      "example": "NO_ERROR"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitExceeded"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/analytics/geography": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "الإحصائيات الجغرافية",
        "description": "توزيع النقرات حسب البلد والموقع الجغرافي",
        "operationId": "getGeographicStats",
        "parameters": [
          {
            "name": "alias",
            "in": "query",
            "description": "الاسم المستعار للرابط المختصر",
            "required": true,
            "schema": {
              "type": "string",
              "example": "myalias123"
            }
          },
          {
            "name": "startDate (UTC)",
            "in": "query",
            "description": "تاريخ البداية (افتراضي: 30 يوماً مضت)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2025-08-01T00:00:00.000Z"
            }
          },
          {
            "name": "endDate (UTC)",
            "in": "query",
            "description": "تاريخ النهاية (افتراضي: الآن)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2028-08-30T23:59:59.000Z"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "تم استرداد الإحصائيات الجغرافية بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "country": {
                            "type": "string",
                            "example": "Egypt"
                          },
                          "count": {
                            "type": "integer",
                            "example": 450
                          },
                          "percentage": {
                            "type": "number",
                            "example": 36.0
                          }
                        }
                      }
                    },
                    "errors": {
                      "type": "array",
                      "items": {},
                      "example": []
                    },
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "errorCode": {
                      "type": "string",
                      "example": "NO_ERROR"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitExceeded"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/analytics/referers": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "إحصائيات المصادر",
        "description": "المواقع والمصادر التي أتت منها النقرات",
        "operationId": "getRefererStats",
        "parameters": [
          {
            "name": "alias",
            "in": "query",
            "description": "الاسم المستعار للرابط المختصر",
            "required": true,
            "schema": {
              "type": "string",
              "example": "myalias123"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "تاريخ البداية (افتراضي: 30 يوماً مضت)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2025-08-01T00:00:00.000Z"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "تاريخ النهاية (افتراضي: الآن)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2025-08-30T23:59:59.000Z"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "تم استرداد إحصائيات المصادر بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "referer": {
                            "type": "string",
                            "example": "facebook.com"
                          },
                          "count": {
                            "type": "integer",
                            "example": 320
                          }
                        }
                      }
                    },
                    "errors": {
                      "type": "array",
                      "items": {},
                      "example": []
                    },
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "errorCode": {
                      "type": "string",
                      "example": "NO_ERROR"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitExceeded"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
     "/analytics/hourly": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "الإحصائيات بالساعة",
        "description": "توزيع النقرات حسب الساعات على مدار اليوم",
        "operationId": "getHourlyStats",
        "parameters": [
          {
            "name": "alias",
            "in": "query",
            "description": "الاسم المستعار للرابط المختصر",
            "required": true,
            "schema": {
              "type": "string",
              "example": "myalias123"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "تاريخ البداية (افتراضي: 7 أيام مضت)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2025-08-24T00:00:00.000Z"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "تاريخ النهاية (افتراضي: الآن)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2025-08-30T23:59:59.000Z"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "تم استرداد الإحصائيات بالساعة بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "hour": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 23,
                            "example": 14
                          },
                          "clicks": {
                            "type": "integer",
                            "example": 85
                          }
                        }
                      }
                    },
                     "errors": {
                      "type": "array",
                      "items": {},
                      "example": []
                    },
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "errorCode": {
                      "type": "string",
                      "example": "NO_ERROR"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitExceeded"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "UnauthorizedError": {
        "description": "غير مسموح لك بالوصول إلى هذا المورد.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "example": {}
                },
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "غير مسموح لك بتنفيذ هذا الإجراء على هذا المورد."
                  ]
                },
                "errorCode": {
                  "type": "string",
                  "example": "UNAUTHORIZED"
                },
                "code": {
                  "type": "number",
                  "example": 2
                }
              }
            }
          }
        }
      },
      "RateLimitExceeded": {
        "description": "عدد كبير من الطلبات — تم تجاوز الحد المسموح.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "example": {}
                },
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "لقد قمت بعدد كبير من الطلبات في وقت قصير. الرجاء الانتظار قليلاً ثم المحاولة مجددًا."
                  ]
                },
                "errorCode": {
                  "type": "string",
                  "example": "RATE_LIMIT_EXCEEDED"
                },
                "code": {
                  "type": "number",
                  "example": 5
                }
              }
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "خطأ داخلي في الخادم.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "example": {}
                },
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "حدث خطأ غير متوقع في الخادم. يرجى المحاولة لاحقًا."
                  ]
                },
                "errorCode": {
                  "type": "string",
                  "example": "INTERNAL_SERVER_ERROR"
                },
                "code": {
                  "type": "number",
                  "example": 4
                }
              }
            }
          }
        }
      },
      "NotFoundError": {
        "description": "المورد المطلوب غير موجود.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "example": {}
                },
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "الرابط المختصر غير موجود"
                  ]
                },
                "errorCode": {
                  "type": "string",
                  "example": "NOT_FOUND"
                },
                "code": {
                  "type": "number",
                  "example": 6
                }
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "رمز Bearer للمصادقة"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}